// packages/server/src/Interface.ts
export type SupportChannel = 'web' | 'feishu' | 'wecom' | 'dingtalk' | 'mp' | 'unknown'